Consolidate rigid joint equality test coverage.
