+++ tests/rigid/test_mujoco_parity.py
@@ -17,11 +17,15 @@ from ..utils.mujoco_parity import (
 
 
 @pytest.mark.required
-@pytest.mark.parametrize("model_name", ["single_joint_equality"])
+@pytest.mark.parametrize("model_name", ["scaled_mjcf_joint_equalities"])
 @pytest.mark.parametrize("gs_solver, gs_integrator", [(gs.constraint_solver.Newton, gs.integrator.implicitfast)])
 @pytest.mark.parametrize("backend", [gs.cpu])
-def test_equality_joint_constant(gs_sim, mj_sim, tol):
-    simulate_and_check_mujoco_consistency(gs_sim, mj_sim, qpos=(0.0, 1.0), num_steps=50, tol=tol)
+def test_equality_joint(gs_sim, mj_sim, tol):
+    (entity,) = gs_sim.entities
+    qpos = entity.get_qpos()
+    (i_unrelated_q,) = entity.get_joint("unrelated").qs_idx_local
+    qpos[i_unrelated_q] = 1.0
+    simulate_and_check_mujoco_consistency(gs_sim, mj_sim, qpos=qpos, num_steps=50, tol=tol)
 
 
 @pytest.mark.required
